Fixed debugger for imm shift instructions

+++ src/mips/instruction.cpp
@@ -65,7 +65,7 @@ namespace mips
                     os << opcode_to_string(instr.op)
                     << " " << reg_to_string(instr.rd)
                     << ", " << reg_to_string(instr.rs)
-                    << ", " << instr.sa;
+                    << ", " << static_cast<uint32_t>(instr.sa);
                 }
                 break;
             }
